+++ src/main/java/edu/ie3/datamodel/io/factory/input/participant/EvcsInputFactory.java
@@ -31,10 +31,10 @@ public class EvcsInputFactory
     extends SystemParticipantInputEntityFactory<EvcsInput, SystemParticipantEntityData> {
 
   private static final String TYPE = "type";
-  private static final String CHARGING_POINTS = "chargingpoints";
-  private static final String COS_PHI_RATED = "cosphirated";
-  private static final String LOCATION_TYPE = "locationtype";
-  private static final String V2G_SUPPORT = "v2gsupport";
+  private static final String CHARGING_POINTS = "chargingPoints";
+  private static final String COS_PHI_RATED = "cosPhiRated";
+  private static final String LOCATION_TYPE = "locationType";
+  private static final String V2G_SUPPORT = "v2gSupport";
 
   public EvcsInputFactory() {
     super(EvcsInput.class);