+++ src/main/java/edu/ie3/datamodel/io/factory/typeinput/Transformer3WTypeInputFactory.java
@@ -17,32 +17,32 @@ import tech.units.indriya.ComparableQuantity;
 
 public class Transformer3WTypeInputFactory
     extends AssetTypeInputEntityFactory<Transformer3WTypeInput> {
-  private static final String S_RATED_A = "srateda";
-  private static final String S_RATED_B = "sratedb";
-  private static final String S_RATED_C = "sratedc";
-  private static final String V_RATED_A = "vrateda";
-  private static final String V_RATED_B = "vratedb";
-  private static final String V_RATED_C = "vratedc";
-  private static final String R_SC_A = "rsca";
-  private static final String R_SC_B = "rscb";
-  private static final String R_SC_C = "rscc";
-  private static final String X_SC_A = "xsca";
-  private static final String X_SC_B = "xscb";
-  private static final String X_SC_C = "xscc";
-  private static final String G_M = "gm";
-  private static final String B_M = "bm";
-  private static final String D_V = "dv";
-  private static final String D_PHI = "dphi";
-  private static final String TAP_NEUTR = "tapneutr";
-  private static final String TAP_MIN = "tapmin";
-  private static final String TAP_MAX = "tapmax";
+  private static final String S_RATED_A = "sRatedA";
+  private static final String S_RATED_B = "sRatedB";
+  private static final String S_RATED_C = "sRatedC";
+  private static final String V_RATED_A = "vRatedA";
+  private static final String V_RATED_B = "vRatedB";
+  private static final String V_RATED_C = "vRatedC";
+  private static final String R_SC_A = "rScA";
+  private static final String R_SC_B = "rScB";
+  private static final String R_SC_C = "rScC";
+  private static final String X_SC_A = "xScA";
+  private static final String X_SC_B = "xScB";
+  private static final String X_SC_C = "xScC";
+  private static final String G_M = "gM";
+  private static final String B_M = "bM";
+  private static final String D_V = "dV";
+  private static final String D_PHI = "dPhi";
+  private static final String TAP_NEUTR = "tapNeutr";
+  private static final String TAP_MIN = "tapMin";
+  private static final String TAP_MAX = "tapMax";
 
   public Transformer3WTypeInputFactory() {
     super(Transformer3WTypeInput.class);
   }
 
   @Override
-  protected List<Set<String>> getFields(EntityData data) {
+  protected List<Set<String>> getFields(Class<?> entityClass) {
     Set<String> constructorParams =
         newSet(
             ENTITY_UUID,