+++ src/main/java/edu/ie3/datamodel/io/source/TypeSource.java
@@ -6,7 +6,9 @@
 package edu.ie3.datamodel.io.source;
 
 import edu.ie3.datamodel.exceptions.FactoryException;
+import edu.ie3.datamodel.exceptions.FailedValidationException;
 import edu.ie3.datamodel.exceptions.SourceException;
+import edu.ie3.datamodel.exceptions.ValidationException;
 import edu.ie3.datamodel.io.factory.EntityData;
 import edu.ie3.datamodel.io.factory.EntityFactory;
 import edu.ie3.datamodel.io.factory.input.OperatorInputFactory;
@@ -22,6 +24,8 @@ import edu.ie3.datamodel.models.input.connector.type.Transformer2WTypeInput;
 import edu.ie3.datamodel.models.input.connector.type.Transformer3WTypeInput;
 import edu.ie3.datamodel.models.input.system.type.*;
 import edu.ie3.datamodel.utils.Try;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.stream.Stream;
@@ -50,6 +54,32 @@ public class TypeSource extends EntitySource {
     this.lineTypeInputFactory = new LineTypeInputFactory();
     this.transformer3WTypeInputFactory = new Transformer3WTypeInputFactory();
     this.systemParticipantTypeInputFactory = new SystemParticipantTypeInputFactory();
+  }
+
+  @Override
+  public void validate() throws ValidationException {
+    List<Try<Void, ValidationException>> participantResults =
+        new ArrayList<>(
+            Stream.of(
+                    EvTypeInput.class,
+                    HpTypeInput.class,
+                    BmTypeInput.class,
+                    WecTypeInput.class,
+                    ChpTypeInput.class,
+                    StorageTypeInput.class)
+                .map(c -> validate(c, systemParticipantTypeInputFactory))
+                .toList());
+
+    participantResults.addAll(
+        List.of(
+            validate(OperatorInput.class, operatorInputFactory),
+            validate(LineTypeInput.class, lineTypeInputFactory),
+            validate(Transformer2WTypeInput.class, transformer2WTypeInputFactory),
+            validate(Transformer3WTypeInput.class, transformer3WTypeInputFactory)));
+
+    Try.scanCollection(participantResults, Void.class)
+        .transformF(FailedValidationException::new)
+        .getOrThrow();
   }
 
   /**