+++ src/main/java/edu/ie3/datamodel/io/source/csv/CsvJointGridContainerSource.java
@@ -8,6 +8,7 @@ package edu.ie3.datamodel.io.source.csv;
 import edu.ie3.datamodel.exceptions.FileException;
 import edu.ie3.datamodel.exceptions.InvalidGridException;
 import edu.ie3.datamodel.exceptions.SourceException;
+import edu.ie3.datamodel.exceptions.ValidationException;
 import edu.ie3.datamodel.io.naming.DefaultDirectoryHierarchy;
 import edu.ie3.datamodel.io.naming.EntityPersistenceNamingStrategy;
 import edu.ie3.datamodel.io.naming.FileNamingStrategy;
@@ -19,9 +20,7 @@ import edu.ie3.datamodel.models.input.connector.type.LineTypeInput;
 import edu.ie3.datamodel.models.input.container.*;
 import edu.ie3.datamodel.utils.Try;
 import java.nio.file.Path;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
+import java.util.*;
 
 /** Convenience class for cases where all used data comes from CSV sources */
 public class CsvJointGridContainerSource {
@@ -55,6 +54,16 @@ public class CsvJointGridContainerSource {
     SystemParticipantSource systemParticipantSource =
         new SystemParticipantSource(typeSource, thermalSource, rawGridSource, emSource, dataSource);
     GraphicSource graphicSource = new GraphicSource(typeSource, rawGridSource, dataSource);
+
+    /* validating sources */
+    try {
+      typeSource.validate();
+      rawGridSource.validate();
+      systemParticipantSource.validate();
+      graphicSource.validate();
+    } catch (ValidationException ve) {
+      throw new SourceException("Could not read source because validation failed", ve);
+    }
 
     /* Loading basic inputs that are used multiple times */
     Map<UUID, OperatorInput> operators = typeSource.getOperators();